Folder & file cleanup. Concept of ECS like to make project more manageable going forward.
