+++ plugins/filament_view/core/entity/entityobject.cc
