+++ plugins/filament_view/core/entity/entityobject.h
