+++ plugins/filament_view/core/entity/model/animation/animation.cc
