+++ plugins/filament_view/core/entity/model/animation/animation_manager.cc
