+++ plugins/filament_view/core/entity/model/loader/model_loader.cc
