+++ plugins/filament_view/core/entity/model/state/shape_state.h
