+++ plugins/filament_view/core/entity/shapes/baseshape.cc
