+++ plugins/filament_view/core/entity/shapes/cube.cc
