+++ plugins/filament_view/core/entity/shapes/sphere.cc
