+++ plugins/filament_view/core/systems/collision_manager.cc
