+++ plugins/filament_view/viewer/custom_model_viewer.h
@@ -35,14 +35,13 @@
 #include <wayland-client.h>
 #include <asio/io_context_strand.hpp>
 
-#include "core/model/loader/model_loader.h"
-#include "core/model/model.h"
-#include "core/model/state/model_state.h"
-#include "core/model/state/scene_state.h"
-#include "core/model/state/shape_state.h"
+#include "core/entity/model/loader/model_loader.h"
+#include "core/entity/model/model.h"
+#include "core/entity/model/state/model_state.h"
+#include "core/entity/model/state/scene_state.h"
+#include "core/entity/model/state/shape_state.h"
 #include "core/scene/camera/camera_manager.h"
 #include "core/scene/scene.h"
-#include "core/shapes/baseshape.h"
 #include "flutter_desktop_plugin_registrar.h"
 #include "platform_views/platform_view.h"
 #include "settings.h"
